pyproject.toml
[project]
name = "marketing-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[google-genai,tools]>=1.7.2",
    "crewai-tools>=1.7.2",
    "python-dotenv>=1.1.1",
    "litellm>=1.75.3",
    "google-generativeai>=0.8.6",
    "streamlit>=1.52.2",
    "openai>=1.83.0",
    "python-docx>=1.2.0",
    "fpdf>=1.7.2",
    "streamlit-authenticator>=0.4.2",
]
